refactor: maintenance-htmlをsrc/client/templatesに移動
src/libに単一ファイルのみ存在していたため、より適切な場所として
src/client/templates配下に移動し、コード構成のシンプルさを維持する。
